dev/requirements.txt
tqdm
kagglehub
torchaudio
soundfile
mlop[full]
torch
torchmetrics
scikit-learn
python-dotenv
fire
encodec